Add daily POTD notifications(safely)
